pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "consolemod"
version = "0.2.1"
description = "A powerful, thread-safe terminal UI library with multi-pane support and async capabilities"
readme = "README.md"

authors = [
    {name = "404Development LLC", email = "consolemod@404development.dev"},
    {name = "Conner Adams"}
]
requires-python = ">=3.9"
dependencies = [
    "rich>=14.2.0",
    "readchar>=4.2.1",
    "pyyaml>=6.0.1"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13"
]

[project.license]
file = "LICENSE"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950"
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/ConnerAdamsMaine/ConsoleMod"
Documentation = "https://github.com/ConnerAdamsMaine/ConsoleMod/blob/main/README.md"
Repository = "https://github.com/ConnerAdamsMaine/ConsoleMod"
"Bug Tracker" = "https://github.com/ConnerAdamsMaine/ConsoleMod/issues"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --strict-markers"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["consolemod*"]
exclude = ["tests*", "Examples*", "Docs*"]